@@ -7,7 +7,7 @@ fun environment(key: String) = providers.environmentVariable(key)
 
 plugins {
     id("org.jetbrains.changelog") version "2.2.0"
-    id("org.jetbrains.intellij.platform") version "2.0.0-beta4"
+    id("org.jetbrains.intellij.platform") version "2.0.1"
     id("java")
     id("org.sonarqube") version "5.0.0.4638"
     id("jacoco")
@@ -26,9 +26,10 @@ repositories {
 
 dependencies {
     implementation("com.google.code.gson:gson:2.11.0")
-    implementation("io.sentry:sentry:7.9.0")
+    implementation("io.sentry:sentry:7.14.0")
 
     val junitVersion = "5.10.2"
+    testImplementation("junit:junit:4.13.2")
     testImplementation("org.junit.jupiter:junit-jupiter-api:$junitVersion")
     testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:$junitVersion")
     testImplementation("org.junit.jupiter:junit-jupiter-params:$junitVersion")
@@ -42,7 +43,7 @@ dependencies {
         pluginVerifier("1.307")
         zipSigner()
         instrumentationTools()
-        testFramework(TestFrameworkType.Platform.JUnit4)
+        testFramework(TestFrameworkType.Platform)
 
         bundledPlugins(
             "com.intellij.database",
@@ -61,16 +62,16 @@ dependencies {
          * @link https://plugins.jetbrains.com/plugin/8595-php-docker/versions
          * */
         plugins(
-            "com.jetbrains.php:241.14494.240",
-            "org.jetbrains.plugins.phpstorm-remote-interpreter:241.14494.158",
-            "org.jetbrains.plugins.phpstorm-docker:241.14494.158"
+            "com.jetbrains.php:242.21829.15",
+            "org.jetbrains.plugins.phpstorm-remote-interpreter:242.20224.159",
+            "org.jetbrains.plugins.phpstorm-docker:242.20224.159"
         )
     }
 }
 
 java {
     toolchain {
-        languageVersion.set(JavaLanguageVersion.of(17))
+        languageVersion.set(JavaLanguageVersion.of(21))
     }
 }
 
@@ -98,7 +99,7 @@ intellijPlatform {
         password.set(environment("PRIVATE_KEY_PASSWORD"))
     }
 
-    verifyPlugin {
+    pluginVerification {
         ignoredProblemsFile = file("ignoredProblems.txt")
         ides {
             ide(IntelliJPlatformType.PhpStorm, properties("platformVersion").get())
